Refactor MUX tests
